refactor(structure);feature(classes)
- Moves functions used in creating the structure from `layopt` to a new module `structure`. This is required because
otherwise we encounter circular arguments when moving to the `Structure` dataclass which uses the `__post_init__()`
method to generate the structure elements on instantiation.
- Introduces the `Structure` dataclass which holds the structures/features.
- Includes tests for new classes and moves tests from `test_layopt.py` to `test_structure.py` where functions have
moved.

Currently `Structure` is _not_ used in the processing done by `trussopt()` introducing that will be a separate piece of
work.
