+++ src/layopt/layopt.py
@@ -1,10 +1,5 @@
 """Layopt module."""
 
-# below line for use in Colab
-# @title { vertical-output: true}
-
-## AF WIP 20260217
-
 # -*- coding: utf-8 -*-
 ## This file forms supplementary material to the paper
 ## "Adaptive topology optimization of fail-safe truss structures" by
@@ -14,11 +9,9 @@
 ## He, L., Gilbert, M. & Song, X. A Python script for adaptive layout
 ## optimization of trusses. Struct Multidisc Optim 60, 835â€“847 (2019).
 ## https://doi.org/10.1007/s00158-019-02226-6
-# !pip install mosek
 
-import itertools
 import time
-from math import ceil, gcd, isinf
+from math import ceil, isinf
 from pathlib import Path
 
 import cvxpy as cvx
@@ -28,14 +21,10 @@ import pandas as pd
 from loguru import logger
 from scipy import sparse
 
-# from numpy.matlib import repmat
-from shapely.geometry import LineString, Point, Polygon
-
+from layopt import structure
 from layopt.classes import Parameters
 from layopt.io import dict_to_df, get_date_time
 from layopt.plotting import plot_truss
-
-# pylint: disable=too-many-lines
 
 
 def calc_eq_matrix_b(
@@ -237,80 +226,6 @@ def stop_violation(
     for i in range(min(num, len(vio_sort))):
         potential_members[lst[vio_c_n[vio_sort[i]]]][3] = True  # set member as active
     return min(num, len(vio_sort))
-
-
-def make_pattern_loads(
-    nodal_coords: npt.NDArray[np.float64],
-    loaded_points: npt.NDArray[np.int64],
-    load_large: float = 50.0,
-    load_small: float = 5.0,
-    load_direction: tuple[float, float] = (0.0, -1.0),
-) -> tuple[list[npt.NDArray[np.float64]], npt.NDArray[np.float64], list[str]]:
-    """
-    Generate all 2^n combinations of large/small loads at each load point.
-
-    Parameters
-    ----------
-    nodal_coords : npt.NDArray[np.float64]
-        Nodal coordinates.
-    loaded_points : npt.NDArray[np.int64]
-        Load points.
-    load_large : float
-        Large load to apply at each load point (default=``50``).
-    load_small : float
-        Small load to apply at each load point (default=``5``).
-    load_direction : tuple[float, float]
-        Load direction (default=``(0,-1)``).
-
-    Returns
-    -------
-    tuple[list[npt.NDArray[np.float64]], npt.NDArray[np.float64], list[str]]
-        A tuple consisting of ``all_patterns`` (all load cases), ``base_load``
-        (base load case) and ``pattern_descriptions`` (description of each load
-        case using ``L`` for large or ``S`` for small at each load point).
-    """
-    if not isinstance(loaded_points, np.ndarray):
-        msg = f"'loaded_points' is not a numpy array : {type(loaded_points)=}"
-        raise TypeError(msg)
-    try:
-        assert loaded_points.shape[1] >= 1, IndexError(
-            f"Need at least one load point : {loaded_points.shape=}"
-        )
-    except IndexError as e:
-        msg = f"Need at least one load point : {loaded_points.shape}"
-        raise IndexError(msg) from e
-
-    # Find node indices for each load point
-    load_node_indices = []
-    for loaded_point in loaded_points:
-        dists = np.linalg.norm(nodal_coords - np.array(loaded_point), axis=1)
-        load_node_indices.append(np.argmin(dists))
-
-    # ns-rse 2026-03-17 : Could maybe use a dictionary here to link patterns and descriptions?
-    all_patterns = []
-    pattern_descriptions = []
-
-    # Generate all 2^n combinations
-    # First combo (all loadLarge) is base case
-    for combo in itertools.product([load_large, load_small], repeat=len(loaded_points)):
-        fk = np.zeros(len(nodal_coords) * 2)
-        desc = []
-        for pt_idx, magnitude in enumerate(combo):
-            node = load_node_indices[pt_idx]
-            fk[node * 2] += magnitude * load_direction[0]
-            fk[node * 2 + 1] += magnitude * load_direction[1]
-            desc.append(f"pt{pt_idx}={'L' if magnitude == load_large else 'S'}")
-
-        all_patterns.append(fk)
-        pattern_descriptions.append(", ".join(desc))
-
-    # ns-rse 2026-03-17 : Return directly as part of tuple
-    base_load = all_patterns[0]  # First pattern = all large loads
-    logger.info(
-        f"Total patterns for {len(loaded_points)} load point(s) : {len(all_patterns)}"
-    )
-    logger.info(f"Base case (all large) : {pattern_descriptions[0]}")
-    return all_patterns, base_load, pattern_descriptions
 
 
 # ns-rse 2026-03-23 : Could this comment perhaps form the extended description for the function?
@@ -589,179 +504,6 @@ def stop_primal_violation_pattern(
     return True  # converged, terminate
 
 
-def _support_conditions(
-    nodal_coords: npt.NDArray[np.float64], support_points: npt.NDArray[np.float64]
-) -> npt.NDArray[np.float64]:
-    """
-    Create the degrees of freedom for support conditions.
-
-    Parameters
-    ----------
-    nodal_coords : npt.NDArray[np.float64]
-        Coordinates for all nodes.
-    support_points : npt.NDArray[np.float64]
-        Preselected support points.
-
-    Returns
-    -------
-    npt.NDArray[np.float64]
-        Flattened array of degrees of freedom.
-    """
-    dof = np.ones((len(nodal_coords), 2))
-    for i, node in enumerate(nodal_coords):
-        if support_points.size == 0:
-            if node[0] == 0:
-                dof[i, :] = [0, 0]  # Support nodes with x=0
-        else:
-            dof[i, :] = (
-                [0, 0]
-                if any((node == point).all() for point in support_points)
-                else [1, 1]
-            )
-    return np.array(dof).flatten()
-
-
-def _potential_members(
-    nodal_coords: npt.NDArray,
-    max_length: float,
-    joint_cost: float,
-    convex: bool,
-    polygon: Polygon,
-    active_member_threshold: float = 1.5,
-) -> npt.NDArray[np.float64]:
-    """
-    Create the ground structure.
-
-    Parameters
-    ----------
-    nodal_coords : npt.NDArray,
-        Node coordinates.
-    max_length : int | float,
-        Maximum length.
-    joint_cost : int | float,
-        Joint cost.
-    convex : bool,
-        Whether the structure is convex.
-    polygon : Polygon
-        Bounding box for the structure.
-    active_member_threshold : float
-        Threshold for marking a potential member as active.
-
-    Returns
-    -------
-    npt.NDArray[np.float64]
-        Array of ground structure coordinates.
-    """
-    potential_members = []
-    for i, j in itertools.combinations(range(len(nodal_coords)), 2):
-        dx, dy = (
-            abs(nodal_coords[i][0] - nodal_coords[j][0]),
-            abs(nodal_coords[i][1] - nodal_coords[j][1]),
-        )
-        length = np.sqrt(dx**2 + dy**2)
-        # Remove overlapping members, or members longer than maxLength
-        if (length < max_length and gcd(int(dx), int(dy)) == 1) or joint_cost != 0:
-            seg = [] if convex else LineString([nodal_coords[i], nodal_coords[j]])
-            if convex or polygon.contains(seg) or polygon.boundary.contains(seg):
-                # Create active members
-                if length <= active_member_threshold:
-                    potential_members.append([i, j, length, True])
-                else:
-                    potential_members.append([i, j, length, False])
-
-    return np.asarray(potential_members)
-
-
-def _primal_adaptivity(
-    primal_method: str, all_patterns_length: int
-) -> tuple[bool, npt.NDArray[np.int32]]:
-    """
-    Derive primal method and active load cases. Start with base load for cases only.
-
-    Parameters
-    ----------
-    primal_method : str
-        Primal method.
-    all_patterns_length : int
-        All patterns.
-
-    Returns
-    -------
-    tuple[bool, npt.NDArray[np.int32]]
-        A tuple of a boolean for ``primal_method`` and the associated ``active_load_cases``.
-    """
-    if primal_method in {"residual", "load_factor"}:
-        active_load_cases = np.zeros(all_patterns_length, dtype=int)
-        active_load_cases[0] = 1  # Base case = all large loads
-        return (True, active_load_cases)
-    return (False, np.ones(all_patterns_length, dtype=int))
-
-
-def _make_polygon(bounding_coordinates: npt.NDArray[np.int32]) -> Polygon:
-    """
-    Construct a ``Polygon`` domain based on the supplied coordinates.
-
-    Traditionally these are rectangular or square but there is no reason that the array of points can not be any other
-    shape.
-
-    Parameters
-    ----------
-    bounding_coordinates : npt.NDArray[np.int32]
-        Coordinates that bound the structure.
-
-    Returns
-    -------
-    Polygon
-        A ``Polygon`` object (from the shapely package).
-    """
-    return Polygon(bounding_coordinates)
-
-
-def _create_nodes(width: int, height: int, polygon: Polygon) -> npt.NDArray[np.float64]:
-    """
-    Create the nodes for the structure.
-
-    Parameters
-    ----------
-    width : int | float
-        The width of the structure.
-    height : int | float
-        The height of the structure.
-    polygon : Polygon
-        Polygon of the structure.
-
-    Returns
-    -------
-    npt.NDArray[np.float64]
-        Two dimensional array of node coordinates.
-    """
-    xv, yv = np.meshgrid(range(width + 1), range(height + 1))
-    points = [Point(xv.flat[i], yv.flat[i]) for i in range(xv.size)]
-    logger.debug(f"Points created : {len(points)=}")
-    return np.array([[pt.x, pt.y] for pt in points if polygon.intersects(pt)])
-
-
-def _loaded_points(width: int, height: int) -> npt.NDArray[np.float64]:
-    """
-    Calculate loaded points based on width and height.
-
-    Loaded points are calculated as being located at the width and mid-point of the height.
-
-    Parameters
-    ----------
-    width : int
-        Width of structure.
-    height : int
-        Height of structure.
-
-    Returns
-    -------
-    npt.NDArray[np.float64]
-        Array of loaded points.
-    """
-    return np.asarray([[width, height // 2]])
-
-
 # Main function - edited for pattern loading
 def trussopt(
     parameters: Parameters,
@@ -792,31 +534,26 @@ def trussopt(
             [0, parameters.height],
         ]
     )
-    poly = _make_polygon(bounding_coordinates)
+    poly = structure.make_polygon(bounding_coordinates)
     convex = poly.convex_hull.area == poly.area
     logger.debug(f"Domain created, convex? : {convex=}")
 
     # Make nodes
-    nodal_coords: npt.NDArray[np.float64] = _create_nodes(
+    nodal_coords: npt.NDArray[np.float64] = structure.create_nodes(
         width=parameters.width, height=parameters.height, polygon=poly
     )
     logger.debug(f"Node coordinates :\n{nodal_coords=}")
-    # ns-rse 2026-05-13 Build a list of nodes to be added to Structure where do loading and virtual_displacements come from?
-    # all_nodes = [
-    #     Node(coordinate=param[0], supported_dof=param[1])
-    #     for param in zip(nodal_coords, dof)
-    # ]
 
     # Default load point
     if parameters.loaded_points is None:
-        parameters.loaded_points = _loaded_points(
+        parameters.loaded_points = structure.calc_default_loaded_points(
             width=parameters.width, height=parameters.height
         )
         logger.info(
             f"Loaded points not provided, calculated as : {parameters.loaded_points=}"
         )
     # Calculate support conditions/degrees of freedom
-    dof = _support_conditions(
+    dof = structure.support_conditions(
         nodal_coords=nodal_coords, support_points=parameters.support_points
     )
     logger.debug(f"Degrees of Freedom : {dof=}")
@@ -824,7 +561,7 @@ def trussopt(
     # ns-rse 2026-03-17 : Unused return arguments but may combine all_patterns and pattern_descriptions to dict
     # all_patterns, base_load, pattern_descriptions = make_pattern_loads(
     # ns-rse 2026-05-13 - loaded_points and load_direction are proposed to be attributes of CaseFamily
-    all_patterns, _, _ = make_pattern_loads(
+    all_patterns, _, _ = structure.make_pattern_loads(
         nodal_coords,
         parameters.loaded_points,
         parameters.load_large,
@@ -833,7 +570,7 @@ def trussopt(
     )
 
     # Create the 'ground structure'
-    potential_members = _potential_members(
+    potential_members = structure.calc_potential_members(
         nodal_coords=nodal_coords,
         max_length=parameters.max_length,
         joint_cost=parameters.joint_cost,
@@ -842,7 +579,7 @@ def trussopt(
         active_member_threshold=parameters.active_member_threshold,
     )
     # Primal adaptivity: start with base load case only ####
-    primal_adaptivity, active_load_cases = _primal_adaptivity(
+    primal_adaptivity, active_load_cases = structure.primal_adaptivity(
         primal_method=parameters.primal_method, all_patterns_length=len(all_patterns)
     )
